requirements.txt
Flask
Werkzeug
pypdf
python-docx
flask-cors